Move StumblerBundleReceiver inside the service.
