+++ src/org/mozilla/mozstumbler/service/StumblerBundleReceiver.java
@@ -1,4 +1,4 @@
-package org.mozilla.mozstumbler.client;
+package org.mozilla.mozstumbler.service;
 
 import android.content.BroadcastReceiver;
 import android.content.ContentValues;
@@ -8,7 +8,7 @@ import android.util.Log;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.mozilla.mozstumbler.StumblerBundle;
+import org.mozilla.mozstumbler.service.StumblerBundle;
 import org.mozilla.mozstumbler.DatabaseContract;
 
 public final class StumblerBundleReceiver extends BroadcastReceiver {
